A part is a pointer, a kind has a name, and part_setup was shadowing seven routines
**`part_setup` had two implementations of nine setups and source order picked
the winner.** An `if` chain dispatched each offset to its transcribed routine;
three table-driven blocks handled groups by walking `flagged[]`, `two[]` and
`sized[]`. They overlapped on ten offsets, and because the table loops came
first, **seven transcribed routines were unreachable** - `part_setup_bellow`,
`part_setup_pokey`, `part_setup_fan`, `part_setup_flashlight`,
`part_setup_scissors`, `part_setup_dynamite_plunger` and `part_setup_08a1`,
each with exactly one call site, in the `if` that never ran. The generalised
table ran in their place.

The original settles it: `part_setup` is **ours**, a stand-in for calling the
far pointer at +0x2a of a kind's record, and the binary has no grouping at all
- each kind's setup is its own routine at its own address. So the tables go and
the dispatcher is a 44-case switch, one routine per offset. Two routines had to
be written because the tables were their only implementation: `part_setup_ramp`
at 172c:2728, transcribed from the disassembly, and `part_setup_corner_pipe` at
172c:377b, lifted out of an inline block that carried no address at all.

Before deleting anything the tables were checked against the routines they were
shadowing: `part_setup_bellow` is exactly `{0x0371, 0x31e0, 0x31b6, 6, 0xffff,
0xffff}`, and `part_setup_dynamite_plunger` - the one entry with the extra
`set_b`/`clear_b` columns - matches `{0x3294, 0x3404, 0x33e6, 4, 0x3416,
0x340a}` down to doing its two byte writes *before* the copy loop.

**`check_solutions` went from 33 to 34.** A level that did not solve now does.

**The 51 part initialisers were a table too**, and the same argument applies.
`part_inits[43]` held six columns standing in for the bodies of forty-three
routines; the constants live as immediates inside the routines and searching
the whole 214,512-byte image for any two of them adjacent as data finds
nothing. They are fifty-one functions now, each carrying its address, and the
form had already cost three defects - two hooks recorded as storing nothing
because the window read was too short, a byte/word column wrong for +0x58, and
three kinds allocating when they should not. The worst it could not report:
the table had **forty-eight** of the fifty-one, and 0x14ca0, 0x14cd9 and
0x14d0a were missing outright.

**A part is a `struct part *`.** Every function in `parts.c` takes one - 92
converted, 53 `uint16_t si = part;` copies retyped and then collapsed, since
the copy was the original's `mov si,[bp+6]` and meant nothing once the
parameter was the pointer. The conversion went down the tree as far as the
rule allows: twenty-five more in `machine.c` and `machine_draw.c`, stopping at
`queue_part`, which files the offset into guest memory with
`DGU16((uint16_t)(si + 2)) = part` and so keeps it. `make_part` answers one and
`free_part` takes one; `dg_off` wraps in `parts.c` went from 162 to 24.

`PART`/`PARTP`, `POINTS`, `POINT_TABLE` and `POINT16_TABLE` lost `volatile`.
The reason the rest of DGROUP has it is the timer thread, and the timer reaches
no part: `timer_callback` to `redraw_cursor` to `draw_cursor`, and
`restage_object_rect` and `save_or_restore_draw_state` under it, have zero
`PART(` between them.

**The shape tables are a struct.** `struct part_shapes` covers DGROUP
0x3182..0x3576 - bounded by `dg_317e` below and `DG3576` above - as 78 members:
58 `byte_pair` outlines, 9 `point16` tables, 8 offset tables and 3 runs nothing
reads. Every extent is measured from its neighbour and the 78 `DG_ASSERT_AT`s
compiling is the proof, because a wrong length would leave a hole and the 1012
bytes would not add up.

**The kinds have names.** Recovered from the manual and the game's own menus -
the game's word winning, so kind 6 is the mouse cage, 12 is pokey, 15 is bob
the fish and 31 is the monkey. 157 routines are named for the kind they serve,
54 `KIND_*` constants replace 209 comparisons, and `burst_kind_19`,
`break_kind_15` and `trigger_kind_6` became `burst_dynamite`,
`break_bob_the_fish` and `trigger_mouse_cage`. The map is written into
`dgroup.h`, because it is the one thing here that is not derivable from the
binary.

`#define` and not an `enum`: a part's `kind` is two bytes and a C enum is an
`int`. This compiler accepts `enum part_kind_id : uint16_t` under `-std=c11`,
which is the reason not to write it - `struct part`'s layout would then rest on
an extension rather than on the original.

Three kinds have no name in either source and are named from what the code does
with them: the gun fires a **bullet**, a burst leaves a **blast** of shreds, a
cut belt leaves two **anchors**. 55, 56 and 57 keep their numbers, and are not
unused - `part_setup_1105` serves 55 and 57 and tells them apart.

**`mul16x16` answers `int32_t`.** It takes signed operands and does the
original's `imul`, then cast its own result to unsigned - and 48 call sites
cast it straight back. The casts were load-bearing: eight `>> 14` sites in
`machine.c` are the original's `sar`, and without a signed operand they would
be logical shifts on a negative product. Return type and casts moved together.

**`word_78` is `next_linked_ptr`**, and only means anything after one of five
routines builds the chain - `collect_carried` and the four `link_*`. Eighteen
other routines only walk it. The walks keep the offset as the loop variable and
take a `struct part *` inside, because the chain ends on offset 0 and `PARTP(0)`
is a real DGROUP address rather than null.

**`tools/dgrules.py` gained `ptr-arg`**: an operation inside an
offset-to-pointer conversion. It was wrong twice before it was right - matching
every macro that reaches `dgroup +` swept in `DG8` and friends and reported 943
sites, nine in ten of them ordinary field reads, and allowing only bare names
still called `BMPP(set.bmp[0x25])` a finding when that fetches a stored offset
rather than computing one. A pointer macro casts and stops where a value
accessor dereferences, and fetching is not computing: 12 sites, every one real.

Coordinates are decimal. Addresses, masks and struct offsets are not - a first
pass turned `DG8(0x3330 + steps)` into `13104 + steps`, a `& 0xfff0` into
`& 65520` and six part-record offsets into decimal before it was caught.

briefing 0/0/0 across flips 210, 230 and 260; **34 of 34 solutions solve**, no
aborts and nothing reaching `not_transcribed`; `make test` and the build clean
throughout, zero warnings under `-Wall -Wextra -Wshadow`.

+++ reconstruct/src/trig.c
@@ -32,9 +32,9 @@ static const int16_t COS_TABLE[0x801];
  * It takes its arguments with `mov bx, sp` and never sets up BP, like the sine
  * and cosine beside it.
  */
-uint32_t mul16x16(int16_t a, int16_t b)
+int32_t mul16x16(int16_t a, int16_t b)
 {
-    return (uint32_t)((int32_t)a * (int32_t)b);
+    return (int32_t)a * (int32_t)b;
 }
 /*
  * 0x2a456
